+++ types-override/http/prefer.d.ts
@@ -1,0 +1,5 @@
+import type { PreferResult } from '../ergo.js';
+
+declare function prefer(): (req: import('node:http').IncomingMessage) => PreferResult;
+
+export default prefer;